requirements.txt
apache-airflow-providers-snowflake==3.2.0
apache-airflow-providers-amazon==5.0.0